Subsystem container demonstration
This commit demonstrates a way to package all subsystems into a single object, which can then be passed to commands.  The advantage is that if a command requires multiple subsystems, only one parameter needs to be passed in.  This improves maintenance and code updates, as adding or changing subsystems used in a command doesn't require changing the constructor signature.